+++ src/Plivo/Resource/TollfreeVerification/AsyncResponse.cs
@@ -1,0 +1,6 @@
+namespace Plivo.Resource.TollfreeVerification
+{
+    public class AsyncResponse : BaseResponse
+    {
+    }
+}