+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [5.37.0](https://github.com/plivo/plivo-dotnet/tree/v5.37.0) (2023-10-31)
+**Feature - TollFree Verification API Support**
+- API support for Create, Update, Get, Delete and List Tollfree Verification.
+- Added New Param `toll_free_sms_verification_id` and `toll_free_sms_verification_order_status `in to the response of the [list all numbers API], [list single number API]
+- Added `toll_free_sms_verification_order_status` filter to AccountPhoneNumber - list all my numbers API.
+
 ## [5.36.0](https://github.com/plivo/plivo-dotnet/tree/v5.36.0) (2023-10-20)
 **Feature - campaign_source field**
 - Added new param `campaign_source` in LIST / GET Campaign APIs